Extract app header as separate layout component
