+++ web/src/app/app/page.tsx
@@ -1,37 +1,25 @@
 'use client';
 
 import { useState, useEffect, useCallback } from 'react';
-import Image from 'next/image';
-import { Button } from '@/components/ui/button';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { Code2, Database, Eye, GitBranch, Play, Download, Share } from 'lucide-react';
-import Link from 'next/link';
-import { ProjectSwitcher } from '@/components/project-switcher';
+import { Code2, Database, Eye, GitBranch } from 'lucide-react';
 import { ChatInterface } from '@/features/chat';
 import { DynamicERDiagram } from '@/components/dynamic-er-diagram';
 import { CodeEditor, FileMap } from '@/features/code-editor';
 import { DatabaseViewer } from '@/components/database-viewer';
 import { AppPreview } from '@/components/app-preview';
 import { ResizableLayout } from '@/components/resizable-layout';
-import { downloadProjectAsZip } from '@/lib/export-utils';
 import { ChangesIndicator } from '@/components/changes-indicator';
 import { playNotificationSound } from '@/utils/notification';
-import {
-  DatabaseComingSoonModal,
-  PreviewComingSoonModal,
-  DeployComingSoonModal,
-  ShareComingSoonModal
-} from '@/components/coming-soon-modal';
+import { DatabaseComingSoonModal, PreviewComingSoonModal } from '@/components/coming-soon-modal';
 import { useCodeStorage } from '@/contexts/CodeStorageContext';
 import { FileContent } from '@/services/codeStorage';
 import { useProjectManager } from '@/contexts/project-manager-context';
 import ProjectBuildingAnimation from '@/components/project-building-animation';
-import { GithubCreateRepo } from '@/features/github';
-import { UserProfileDropdown } from '@/features/auth';
+import AppHeader from '@/layout/app-header';
 
 export default function AppPage() {
   const [activeTab, setActiveTab] = useState('er-diagram');
-  const [currentProject, setCurrentProject] = useState('Task Manager App');
   const [initialPrompt, setInitialPrompt] = useState<string | undefined>();
   const [generatedFiles, setGeneratedFiles] = useState<FileMap>({});
   const [currentDbml, setCurrentDbml] = useState<string | undefined>();
@@ -131,9 +119,6 @@ export default function AppPage() {
   // Coming Soon Modal states
   const [showDatabaseModal, setShowDatabaseModal] = useState(false);
   const [showPreviewModal, setShowPreviewModal] = useState(false);
-  const [showDeployModal, setShowDeployModal] = useState(false);
-  const [showShareModal, setShowShareModal] = useState(false);
-  const [isExporting, setIsExporting] = useState(false);
 
   useEffect(() => {
     // Get parameters from URL (passed from home page)
@@ -153,25 +138,6 @@ export default function AppPage() {
       window.history.replaceState({}, '', window.location.pathname);
     }
   }, []);
-
-  const handleExport = async () => {
-    try {
-      setIsExporting(true);
-
-      // Check if there are files to export
-      if (!generatedFiles || Object.keys(generatedFiles).length === 0) {
-        alert('No files to export. Please generate some code first by chatting with the AI.');
-        return;
-      }
-
-      await downloadProjectAsZip(generatedFiles, currentProject);
-    } catch (error) {
-      console.error('Export failed:', error);
-      alert('Export failed. Please try again.');
-    } finally {
-      setIsExporting(false);
-    }
-  };
 
   const renderWorkspacePanelLoading = () => (
     <div className="flex-1 flex flex-col items-center justify-center">
@@ -251,54 +217,7 @@ export default function AppPage() {
   return (
     <div className="h-screen bg-slate-50 dark:bg-slate-950 flex flex-col">
       {/* Top Bar */}
-      <header className="border-b border-slate-200 dark:border-slate-800 bg-white dark:bg-slate-900 px-4 py-3 flex items-center justify-between">
-        <div className="flex items-center space-x-4">
-          <div className="flex items-center space-x-2">
-            <Link href="/" className="hover:opacity-80 transition-opacity cursor-pointer">
-              <span className="flex items-center text-2xl font-bold bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent">
-                <Image src="/logo.png" width={40} height={40} alt="Scafoldr logo" />
-                scafoldr
-              </span>
-            </Link>
-          </div>
-
-          <div className="h-6 w-px bg-slate-300 dark:bg-slate-700" />
-
-          <ProjectSwitcher currentProject={currentProject} onProjectChange={setCurrentProject} />
-        </div>
-
-        <div className="flex items-center space-x-3">
-          <GithubCreateRepo activeProjectId={activeProjectId} />
-          <Button
-            variant="outline"
-            size="sm"
-            className="hidden md:flex bg-transparent"
-            onClick={() => setShowShareModal(true)}>
-            <Share className="w-4 h-4 mr-2" />
-            Share
-          </Button>
-          <Button
-            variant="outline"
-            size="sm"
-            className="hidden md:flex bg-transparent"
-            onClick={handleExport}
-            disabled={isExporting}>
-            <Download className="w-4 h-4 mr-2" />
-            {isExporting ? 'Exporting...' : 'Export'}
-          </Button>
-          <Button
-            size="sm"
-            className="bg-gradient-to-r from-blue-500 to-purple-600 hover:from-blue-600 hover:to-purple-700 text-white hover:text-white"
-            onClick={() => setShowDeployModal(true)}>
-            <Play className="w-4 h-4 mr-2" />
-            Deploy
-          </Button>
-
-          <div className="h-6 w-px bg-slate-300 dark:bg-slate-700" />
-
-          <UserProfileDropdown />
-        </div>
-      </header>
+      <AppHeader activeProjectId={activeProjectId} generatedFiles={generatedFiles} />
 
       {/* Main Content with Resizable Layout */}
       <ResizableLayout
@@ -323,18 +242,6 @@ export default function AppPage() {
       <PreviewComingSoonModal
         open={showPreviewModal}
         onOpenChange={setShowPreviewModal}
-        githubRepo="https://github.com/scafoldr/scafoldr"
-      />
-
-      <DeployComingSoonModal
-        open={showDeployModal}
-        onOpenChange={setShowDeployModal}
-        githubRepo="https://github.com/scafoldr/scafoldr"
-      />
-
-      <ShareComingSoonModal
-        open={showShareModal}
-        onOpenChange={setShowShareModal}
         githubRepo="https://github.com/scafoldr/scafoldr"
       />
     </div>
